requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pathspec==0.12.1
PyYAML==6.0.2
setuptools==75.8.2
Werkzeug==3.1.3
